+++ routes/utils.py
@@ -1,0 +1,6 @@
+import quart
+import os
+import requests
+
+app=quart.Quart("")
+app.secret_key = os.environ.get("session")